// datavaerese_frontend_&_backend/types/enum.ts
export enum Module {
    EPIC = "1",
    Project = "2",
    Session = "3",
    Label = "4",
    // Taxonomy = "5",
    StructureGroup = "6",
    UserGroup = "7",
    JSON = "8",
    SessionCodes = "9",
    UserRoles = "10",
    Users = "11",
    Annotation = "12",
    SessionLabel = "13"
}
  
export enum Action {
    CREATE = "1",
    READ = "2",
    UPDATE = "3",
    DELETE = "4",
    Analyse = "5",
    UploadJSON = "6",
    DownloadJSON = "7",
    DownloadAsVersion = "8",
    VIEWALL = "9",
    Lock = "10",
    Export = "13"
}